% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.polygonstyler-package.R
\docType{package}
\name{leaflet.polygonstyler-package}
\alias{leaflet.polygonstyler}
\alias{leaflet.polygonstyler-package}
\title{leaflet.polygonstyler: Leafler.polygonstyler}
\description{
change color of polygon without reload geometries.
}
\author{
\strong{Maintainer}: Lucas Longour \email{lucas.longour@ird.fr}

}
\keyword{internal}
