% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setshapelabel.R
\name{setShapeLabel}
\alias{setShapeLabel}
\title{setShapeLabel}
\usage{
setShapeLabel(
  map,
  data = getMapData(map),
  layerId,
  label = NULL,
  options = NULL
)
}
\arguments{
\item{map}{map object}

\item{data}{data}

\item{layerId}{layerId}

\item{label}{label}

\item{options}{options}
}
\value{
leaflet map object
}
\description{
setShapeLabel
}
