% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setshapestyle.R
\name{setShapeStyle}
\alias{setShapeStyle}
\title{setShapeStyle}
\usage{
setShapeStyle(
  map,
  data = getMapData(map),
  layerId,
  stroke = NULL,
  color = NULL,
  weight = NULL,
  opacity = NULL,
  fill = NULL,
  fillColor = NULL,
  fillOpacity = NULL,
  dashArray = NULL,
  smoothFactor = NULL,
  noClip = NULL,
  options = NULL,
  label = NULL
)
}
\arguments{
\item{map}{map object}

\item{data}{data}

\item{layerId}{layerId}

\item{stroke}{stroke}

\item{color}{color}

\item{weight}{weight}

\item{opacity}{opacity}

\item{fill}{fill}

\item{fillColor}{fill color}

\item{fillOpacity}{fill opacity}

\item{dashArray}{dash array}

\item{smoothFactor}{smooth factor}

\item{noClip}{no clip}

\item{options}{options}

\item{label}{label}
}
\description{
setShapeStyle
}
